Simulate diffusion using fixed-step vectors
In `_drift_charge!`, diffusion was modelled by adding an effective field vector to
the external electric field and the electric field given by
self-repulsion. For small Δt, the effective field vectors for
diffusion would be very large, resulting in saturated behaviour
in the drift velocity calculation. As the saturation is heavily
dependent on the choice of Δt, diffusion is now modelled by
adding fixed-step vectors after calculating the step vectors
from the physical electric fields.